pep621: refactor StandardMetadata as a dataclass
This removes the need requirement for pyproject.toml parsing, as it will
still be useful without it. We are going to allow conversion to Core
Metadata, so being able to build the metadata instance manually will be
useful.

This patch also changes the interface to use all available pypa/packaging
objects, making it more coherent.

Signed-off-by: Filipe Laíns <lains@riseup.net>