setup: Revert bump of minimum requires packaging version
The minimum required version of packaging was bumped just to make the
exception messages exactly predictable for matching them in the tests.
As the tests have been fixed to be more lenient toward the changes in
the wording of the messages, we can revert this change.

Better be more conservative with the required minimum versions of base
packages. In particular, this allows pyproject-metadata to be
installed in Alpine where packaging is a system package and has not
been yet upgraded to the latest release.